+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/crypto/tasks/RedactEventTask.kt
@@ -15,7 +15,7 @@
  */
 package org.matrix.android.sdk.internal.crypto.tasks
 
-import org.greenrobot.eventbus.EventBus
+import org.matrix.android.sdk.internal.network.GlobalErrorReceiver
 import org.matrix.android.sdk.internal.network.executeRequest
 import org.matrix.android.sdk.internal.session.room.RoomAPI
 import org.matrix.android.sdk.internal.session.room.send.SendResponse
@@ -33,10 +33,10 @@ internal interface RedactEventTask : Task<RedactEventTask.Params, String> {
 
 internal class DefaultRedactEventTask @Inject constructor(
         private val roomAPI: RoomAPI,
-        private val eventBus: EventBus) : RedactEventTask {
+        private val globalErrorReceiver: GlobalErrorReceiver) : RedactEventTask {
 
     override suspend fun execute(params: RedactEventTask.Params): String {
-        val executeRequest = executeRequest<SendResponse>(eventBus) {
+        val executeRequest = executeRequest<SendResponse>(globalErrorReceiver) {
             apiCall = roomAPI.redactEvent(
                     txId = params.txID,
                     roomId = params.roomId,