+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/call/GetTurnServerTask.kt
@@ -17,9 +17,9 @@
 package org.matrix.android.sdk.internal.session.call
 
 import org.matrix.android.sdk.api.session.call.TurnServerResponse
+import org.matrix.android.sdk.internal.network.GlobalErrorReceiver
 import org.matrix.android.sdk.internal.network.executeRequest
 import org.matrix.android.sdk.internal.task.Task
-import org.greenrobot.eventbus.EventBus
 import javax.inject.Inject
 
 internal abstract class GetTurnServerTask : Task<GetTurnServerTask.Params, TurnServerResponse> {
@@ -27,10 +27,10 @@ internal abstract class GetTurnServerTask : Task<GetTurnServerTask.Params, TurnS
 }
 
 internal class DefaultGetTurnServerTask @Inject constructor(private val voipAPI: VoipApi,
-                                                            private val eventBus: EventBus) : GetTurnServerTask() {
+                                                            private val globalErrorReceiver: GlobalErrorReceiver) : GetTurnServerTask() {
 
     override suspend fun execute(params: Params): TurnServerResponse {
-        return executeRequest(eventBus) {
+        return executeRequest(globalErrorReceiver) {
             apiCall = voipAPI.getTurnServer()
         }
     }