+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/directory/GetRoomDirectoryVisibilityTask.kt
@@ -16,8 +16,8 @@
 
 package org.matrix.android.sdk.internal.session.room.directory
 
-import org.greenrobot.eventbus.EventBus
 import org.matrix.android.sdk.api.session.room.model.RoomDirectoryVisibility
+import org.matrix.android.sdk.internal.network.GlobalErrorReceiver
 import org.matrix.android.sdk.internal.network.executeRequest
 import org.matrix.android.sdk.internal.session.directory.DirectoryAPI
 import org.matrix.android.sdk.internal.session.directory.RoomDirectoryVisibilityJson
@@ -32,11 +32,11 @@ internal interface GetRoomDirectoryVisibilityTask : Task<GetRoomDirectoryVisibil
 
 internal class DefaultGetRoomDirectoryVisibilityTask @Inject constructor(
         private val directoryAPI: DirectoryAPI,
-        private val eventBus: EventBus
+        private val globalErrorReceiver: GlobalErrorReceiver
 ) : GetRoomDirectoryVisibilityTask {
 
     override suspend fun execute(params: GetRoomDirectoryVisibilityTask.Params): RoomDirectoryVisibility {
-        return executeRequest<RoomDirectoryVisibilityJson>(eventBus) {
+        return executeRequest<RoomDirectoryVisibilityJson>(globalErrorReceiver) {
             apiCall = directoryAPI.getRoomDirectoryVisibility(params.roomId)
         }
                 .visibility