+++ CHANGES.md
@@ -21,7 +21,7 @@ SDK API changes ⚠️:
  -
 
 Build 🧱:
- -
+ - Remove dependency to org.greenrobot.eventbus library
 
 Test:
  -